truthlens/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.0
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
trafilatura==1.12.0
google-genai==1.5.0
groq==0.11.0
pydantic==2.7.0
python-multipart==0.0.9
duckduckgo-search==6.2.6
aiohttp==3.9.5
